# dist/api-docs.json
{
  "swagger": "2.0",
  "basePath": "/api",
...
  "consumes": ["application/json", "multipart/form-data"],
  "produces": ["application/json"],
  "paths": {
...
    "/customers?pageIndex={pageIndex}&pageSize={pageSize}&{propertyName}&{propertyValue}": {
      "get": {
        "tags": ["Customer"],
        "description": "Get All Customers",
        "operationId": "for the URL",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "bearer token",
            "required": true
          },
          {
            "name": "pageIndex",
            "in": "query",
            "description": "PageIndex",
            "required": true,
            "type": "integer"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "PageSize",
            "required": true,
            "type": "integer"
          },
          {
            "name": "propertyName",
            "in": "query",
            "enum": [
              "name",
              "customerType",
              "buildingName",
              "city",
              "province",
              "areaCode",
              "phoneNumber",
              "primaryContactPerson",
              "primaryContactNumber",
              "secondaryContactPerson",
              "secondaryContactNumber",
              "address"
            ],
            "description": "PropertyName",
            "required": false,
            "type": "string"
          },
          {
            "name": "propertyValue",
            "in": "query",
            "description": "propertyValue",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer response",
            "schema": {
              "$ref": "#/definitions/customerResponseWithPagination"
            }
          }
        }
      }
    },
    "/customers/all?{propertyName}&{propertyValue}": {
      "get": {
        "tags": ["Customer"],
        "description": "Get All Customer ",
        "operationId": "for the URL",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "bearer token",
            "required": true
          },
          {
            "name": "propertyName",
            "in": "query",
            "enum": [
              "name",
              "customerType",
              "buildingName",
              "city",
              "province",
              "areaCode",
              "phoneNumber",
              "primaryContactPerson",
              "primaryContactNumber",
              "secondaryContactPerson",
              "secondaryContactNumber",
              "address"
            ],
            "description": "PropertyName",
            "required": false,
            "type": "string"
          },
          {
            "name": "propertyValue",
            "in": "query",
            "description": "propertyValue",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer response",
            "schema": {
              "$ref": "#/definitions/customerResponse"
            }
          }
        }
      }
    },
    "/customers": {
      "post": {
        "tags": ["Customer"],
        "description": "Create new Customer",
        "operationId": "for the URL",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "bearer token",
            "required": true
          },
          {
            "name": "Job Parameters",
            "in": "body",
            "description": "Customer",
            "required": true,
            "schema": {
              "$ref": "#/definitions/customer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer response",
            "schema": {
              "$ref": "#/definitions/customerData"
            }
          }
        }
      }
    },
    "/customers/{id}/details": {
      "get": {
        "tags": ["Customer"],
        "description": "get by id ",
        "operationId": "for the URL",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "bearer token",
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer response",
            "schema": {
              "$ref": "#/definitions/customerData"
            }
          }
        }
      }
    },
    "/customers/{id}": {
      "put": {
        "tags": ["Customer"],
        "description": "Update Customer",
        "operationId": "for the URL",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "bearer token",
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Job Parameters",
            "in": "body",
            "description": "Customer",
            "required": true,
            "schema": {
              "$ref": "#/definitions/customer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer response",
            "schema": {
              "$ref": "#/definitions/successMessageResponse"
            }
          }
        }
      }
    },
    "/customers/{id}/outlets": {
      "put": {
        "tags": ["Customer"],
        "description": "Assign outlets",
        "operationId": "for the URL",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "bearer token",
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Job Parameters",
            "in": "body",
            "description": "Customer",
            "required": true,
            "schema": {
              "$ref": "#/definitions/customerBranch"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer response",
            "schema": {
              "$ref": "#/definitions/successMessageResponse"
            }
          }
        }
      }
    },
    "/customers/{customerId}/all": {
      "get": {
        "tags": ["Customer"],
        "description": "get all Outlets by customerId",
        "operationId": "for the URL",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "bearer token",
            "required": true
          },
          {
            "name": "customerId",
            "in": "path",
            "description": "customerId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/categoryresponse"
              }
            }
          }
        }
      }
    },
    "/customers/{coustmerId}": {
      "delete": {
        "tags": ["Customer"],
        "description": "delete the Customer by id ",
        "operationId": "for the URL",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "bearer token",
            "required": true
          },
          {
            "name": "coustmerId",
            "in": "path",
            "description": "coustmerId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Ok"
          }
        }
      }
    },
...
  },
  "definitions": {
    "successMessageResponse": {
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "customer": {
      "properties": {
        "name": {
          "required": true,
          "type": "string",
          "example": "techno"
        },
        "warehouseId": {
          "required": false,
          "type": "array",
          "example": [1]
        },
        "customerType": {
          "required": true,
          "type": "string",
          "example": "customerType"
        },
        "buildingName": {
          "required": true,
          "type": "string",
          "example": "buildingName"
        },
        "city": {
          "required": true,
          "type": "string",
          "example": "city"
        },
        "province": {
          "required": true,
          "type": "string",
          "example": "province"
        },
        "areaCode": {
          "required": true,
          "type": "string",
          "example": "areaCode"
        },
        "phoneNumber": {
          "required": true,
          "type": "string",
          "example": "phoneNumber"
        },
        "primaryContactPerson": {
          "required": true,
          "type": "string",
          "example": "primaryContactPerson"
        },
        "primaryContactNumber": {
          "required": true,
          "type": "string",
          "example": "primaryContactNumber"
        },
        "primaryEmail": {
          "required": true,
          "type": "string",
          "example": "inturi99@gmail.com"
        },
        "secondaryContactPerson": {
          "required": true,
          "type": "string",
          "example": "secondaryContactPerson"
        },
        "secondaryContactNumber": {
          "required": true,
          "type": "string",
          "example": "secondaryContactNumber"
        },
        "secondaryEmail": {
          "required": true,
          "type": "string",
          "example": "krishna@gmail.com"
        },
        "address": {
          "required": true,
          "type": "string",
          "example": "address"
        },
        "creditLimit": {
          "required": true,
          "type": "number",
          "example": 200.0
        },
        "description": {
          "required": true,
          "type": "string",
          "example": "pharmacy"
        },
        "longitude": {
          "required": true,
          "type": "number",
          "example": 10.6758
        },
        "latitude": {
          "required": true,
          "type": "number",
          "example": -122.3445
        }
      }
    },
    "customerBranch": {
      "properties": {
        "warehouseId": {
          "required": true,
          "type": "array",
          "example": [1, 2]
        }
      }
    },
    "customerData": {
      "properties": {
        "id": {
          "required": true,
          "type": "number",
          "example": 1
        },
        "name": {
          "required": true,
          "type": "string",
          "example": "techno"
        },
        "warehouseId": {
          "required": false,
          "type": "array",
          "example": [1]
        },
        "warehouseName": {
          "required": true,
          "type": "string",
          "example": null
        },
        "customerType": {
          "required": true,
          "type": "string",
          "example": "customerType"
        },
        "buildingName": {
          "required": true,
          "type": "string",
          "example": "buildingName"
        },
        "city": {
          "required": true,
          "type": "string",
          "example": "city"
        },
        "province": {
          "required": true,
          "type": "string",
          "example": "province"
        },
        "areaCode": {
          "required": true,
          "type": "string",
          "example": "areaCode"
        },
        "phoneNumber": {
          "required": true,
          "type": "string",
          "example": "phoneNumber"
        },
        "primaryContactPerson": {
          "required": true,
          "type": "string",
          "example": "primaryContactPerson"
        },
        "primaryContactNumber": {
          "required": true,
          "type": "string",
          "example": "primaryContactNumber"
        },
        "primaryEmail": {
          "required": true,
          "type": "string",
          "example": "inturi99@gmail.com"
        },
        "secondaryContactPerson": {
          "required": true,
          "type": "string",
          "example": "secondaryContactPerson"
        },
        "secondaryContactNumber": {
          "required": true,
          "type": "string",
          "example": "secondaryContactNumber"
        },
        "secondaryEmail": {
          "required": true,
          "type": "string",
          "example": "krishna@gmail.com"
        },
        "address": {
          "required": true,
          "type": "string",
          "example": "address"
        },
        "creditLimit": {
          "required": true,
          "type": "number",
          "example": 200.0
        },
        "description": {
          "required": true,
          "type": "string",
          "example": "pharmacy"
        },
        "longitude": {
          "required": true,
          "type": "number",
          "example": 10.6758
        },
        "latitude": {
          "required": true,
          "type": "number",
          "example": -122.3445
        },
        "updatedBy": {
          "required": true,
          "type": "number",
          "example": 1
        },
        "createdAt": {
          "required": true,
          "type": "string",
          "example": "2019-02-02T10:18:16.397Z"
        }
      }
    },
    "customerResponseWithPagination": {
      "properties": {
        "count": {
          "required": true,
          "type": "integer",
          "example": 1
        },
        "rows": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/customerData"
          }
        }
      }
    },
    "customerResponse": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/customerData"
      }
    },
    "categoryresponse": {
      "properties": {
        "id": {
          "required": true,
          "type": "integer",
          "example": 1
        },
        "name": {
          "required": true,
          "type": "string",
          "example": ""
        }
      }
    },
...
  }
}